Add some logging and timing to link checker process
We want to get some more data about how long various parts of the link
checking process are taking, and when they are triggered.  We use statsd
timing to get the length of time, and we log some start/end messages to
get when things are happening.  It may not be enough, but hopefully it'll
tell us something.